+++ src/components/forms/FormFieldsEditor.tsx
@@ -2,15 +2,17 @@ import { useEffect, useState, useRef } from 'react';
 import { Form, TextFormField, SelectFormField } from 'types';
 import { FormFieldType } from 'types/Enums';
 import { useUpdateForm, useFormSubmissions } from 'hooks/Form';
-import { useSnackbar } from 'hooks/Snackbar';
-import { ClickAwayListener, Grow, Paper, Popper, Typography, MenuItem, MenuList, Button, Stack } from '@mui/material';
+import { ClickAwayListener, Grow, Paper, Popper, Typography, MenuItem, MenuList, Button, Stack, TextField } from '@mui/material';
 import FieldEditor from 'components/forms/FieldEditor';
+import { useSnackbar } from 'hooks/Snackbar';
 
 export type FormFieldsEditorProps = {
   form: Form;
+  onSave?: () => void;
+  canEditTitle?: boolean;
 };
 
-const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
+const FormFieldsEditor = ({ form, onSave, canEditTitle }: FormFieldsEditorProps) => {
   const { data: submissions, isLoading: isSubmissionsLoading } = useFormSubmissions(form.id, 1);
   const updateForm = useUpdateForm(form.id);
   const disabledFromSubmissions = (submissions ? Boolean(submissions.count) : true) && !isSubmissionsLoading;
@@ -19,6 +21,7 @@ const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
   const [fields, setFields] = useState<Array<TextFormField | SelectFormField>>(form.fields);
   const [addButtonOpen, setAddButtonOpen] = useState(false);
   const buttonAnchorRef = useRef(null);
+  const [formTitle, setFormTitle] = useState(form.title);
 
   useEffect(() => setFields(form.fields), [form]);
 
@@ -67,10 +70,13 @@ const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
       return;
     }
     updateForm.mutate(
-      { fields: fields, resource_type: form.resource_type },
+      { title: formTitle, fields: fields, resource_type: form.resource_type },
       {
         onSuccess: () => {
           showSnackbar('Spørsmålene ble oppdatert', 'success');
+          if (onSave) {
+            onSave();
+          }
         },
         onError: (e) => {
           showSnackbar(e.detail, 'error');
@@ -82,6 +88,7 @@ const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
   return (
     <>
       <Stack gap={1}>
+        {canEditTitle && <TextField label='Malen sitt navn' onChange={(e) => setFormTitle(e.target.value)} type='text' value={formTitle} />}
         {disabledFromSubmissions && (
           <Typography gutterBottom variant='body2'>
             Du kan ikke endre spørsmålene etter at noen har svart på dem