+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ## Neste versjon
 
+- ✨ **Skjemamaler**. Administratorer kan nå lage maler som man kan ta utgangspunkt i når man lager et nytt skjema.
+
 ## Versjon 1.4.2 (31.01.2022)
 
 - 🦟 **Eventadmin**. Viser at et arrangement er stengt i adminpanelet.