+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 
 - ✨ **Bruker**. Det er mulig å eksportere sin brukerdata.
 - ✨ **Bruker**. Det er mulig å slette sin egen bruker.
+- ✨ **Skjemamaler**. Administratorer kan nå lage maler som man kan ta utgangspunkt i når man lager et nytt skjema.
+
+## Versjon 1.4.2 (31.01.2022)
+
 - 🦟 **Eventadmin**. Viser at et arrangement er stengt i adminpanelet.
 - 🦟 **Stories**. Annonser som har fortløpende opptak vises med "fortløpende opptak".